move device-connect-view to views folder. update paths on several places
