+++ src/electron/flux/types/device-store-data.ts
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DeviceConnectState } from '../../device-connect-view/components/device-connect-state';
+import { DeviceConnectState } from '../../views/device-connect-view/components/device-connect-state';
 
 export interface DeviceStoreData {
     deviceConnectState: DeviceConnectState;
