+++ src/electron/views/device-connect-view/components/device-connect-body.scss
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-@import '../../../common/styles/fonts.scss';
+@import '../../../../common/styles/fonts.scss';
 
 .device-connect-body {
     margin: 16px 32px;
