+++ src/electron/views/device-connect-view/components/device-connect-body.tsx
