+++ src/electron/views/device-connect-view/components/device-connect-connected-device.scss
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-@import '../../../common/styles/colors.scss';
-@import '../../../common/styles/fonts.scss';
+@import '../../../../common/styles/colors.scss';
+@import '../../../../common/styles/fonts.scss';
 
 .device-connect-connected-device {
     .device-connect-spinner {
