+++ src/electron/views/device-connect-view/components/device-connect-connected-device.tsx
@@ -3,7 +3,7 @@
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
 import { Spinner, SpinnerSize } from 'office-ui-fabric-react/lib/Spinner';
 import * as React from 'react';
-import { NamedFC } from '../../../common/react/named-fc';
+import { NamedFC } from '../../../../common/react/named-fc';
 import {
     connectionErrorIcon,
     deviceConnectConnectedDevice,
