+++ src/electron/views/device-connect-view/components/device-connect-port-entry.scss
