+++ src/electron/views/device-connect-view/components/device-connect-port-entry.tsx
@@ -4,7 +4,7 @@ import { DefaultButton, PrimaryButton } from 'office-ui-fabric-react/lib/Button'
 import { MaskedTextField } from 'office-ui-fabric-react/lib/TextField';
 import * as React from 'react';
 
-import { DeviceConnectActionCreator } from '../../flux/action-creator/device-connect-action-creator';
+import { DeviceConnectActionCreator } from '../../../flux/action-creator/device-connect-action-creator';
 import { deviceConnectPortEntry, portNumberField } from './device-connect-port-entry.scss';
 import { DeviceConnectState } from './device-connect-state';
