+++ src/electron/views/device-connect-view/components/electron-external-link.tsx
@@ -3,7 +3,7 @@
 import { Shell } from 'electron';
 import { Link } from 'office-ui-fabric-react/lib/Link';
 import * as React from 'react';
-import { NamedFC } from '../../../common/react/named-fc';
+import { NamedFC } from '../../../../common/react/named-fc';
 
 export interface ElectronExternalLinkProps {
     href: string;
