+++ src/electron/views/device-connect-view/components/telemetry-permission-dialog.scss
