+++ src/electron/views/device-connect-view/components/window-title.tsx
@@ -1,7 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as React from 'react';
-import { NamedFC } from '../../../common/react/named-fc';
+import { NamedFC } from '../../../../common/react/named-fc';
 import { headerText, windowTitle } from './window-title.scss';
 
 export interface WindowTitleProps {
