+++ src/electron/views/device-connect-view/device-connect-view-initializer.ts
@@ -5,27 +5,27 @@ import axios from 'axios';
 import { remote } from 'electron';
 import { createFetchScanResults } from 'electron/platform/android/fetch-scan-results';
 import * as ReactDOM from 'react-dom';
-import { UserConfigurationActions } from '../../background/actions/user-configuration-actions';
-import { getPersistedData, PersistedData } from '../../background/get-persisted-data';
-import { UserConfigurationActionCreator } from '../../background/global-action-creators/user-configuration-action-creator';
-import { IndexedDBDataKeys } from '../../background/IndexedDBDataKeys';
-import { InstallationData } from '../../background/installation-data';
-import { UserConfigurationStore } from '../../background/stores/global/user-configuration-store';
-import { getTelemetryClient } from '../../background/telemetry/telemetry-client-provider';
-import { TelemetryEventHandler } from '../../background/telemetry/telemetry-event-handler';
-import { TelemetryLogger } from '../../background/telemetry/telemetry-logger';
-import { TelemetryStateListener } from '../../background/telemetry/telemetry-state-listener';
-import { initializeFabricIcons } from '../../common/fabric-icons';
-import { getIndexedDBStore } from '../../common/indexedDB/get-indexeddb-store';
-import { IndexedDBAPI, IndexedDBUtil } from '../../common/indexedDB/indexedDB';
-import { BaseClientStoresHub } from '../../common/stores/base-client-stores-hub';
-import { telemetryAppTitle } from '../../content/strings/application';
-import { ElectronAppDataAdapter } from '../adapters/electron-app-data-adapter';
-import { ElectronStorageAdapter } from '../adapters/electron-storage-adapter';
-import { RiggedFeatureFlagChecker } from '../common/rigged-feature-flag-checker';
-import { DeviceConnectActionCreator } from '../flux/action-creator/device-connect-action-creator';
-import { DeviceActions } from '../flux/action/device-actions';
-import { DeviceStore } from '../flux/store/device-store';
+import { UserConfigurationActions } from '../../../background/actions/user-configuration-actions';
+import { getPersistedData, PersistedData } from '../../../background/get-persisted-data';
+import { UserConfigurationActionCreator } from '../../../background/global-action-creators/user-configuration-action-creator';
+import { IndexedDBDataKeys } from '../../../background/IndexedDBDataKeys';
+import { InstallationData } from '../../../background/installation-data';
+import { UserConfigurationStore } from '../../../background/stores/global/user-configuration-store';
+import { getTelemetryClient } from '../../../background/telemetry/telemetry-client-provider';
+import { TelemetryEventHandler } from '../../../background/telemetry/telemetry-event-handler';
+import { TelemetryLogger } from '../../../background/telemetry/telemetry-logger';
+import { TelemetryStateListener } from '../../../background/telemetry/telemetry-state-listener';
+import { initializeFabricIcons } from '../../../common/fabric-icons';
+import { getIndexedDBStore } from '../../../common/indexedDB/get-indexeddb-store';
+import { IndexedDBAPI, IndexedDBUtil } from '../../../common/indexedDB/indexedDB';
+import { BaseClientStoresHub } from '../../../common/stores/base-client-stores-hub';
+import { telemetryAppTitle } from '../../../content/strings/application';
+import { ElectronAppDataAdapter } from '../../adapters/electron-app-data-adapter';
+import { ElectronStorageAdapter } from '../../adapters/electron-storage-adapter';
+import { RiggedFeatureFlagChecker } from '../../common/rigged-feature-flag-checker';
+import { DeviceConnectActionCreator } from '../../flux/action-creator/device-connect-action-creator';
+import { DeviceActions } from '../../flux/action/device-actions';
+import { DeviceStore } from '../../flux/store/device-store';
 import { DeviceConnectViewContainerState } from './components/device-connect-view-container';
 import { ElectronLink } from './components/electron-link';
 import { DeviceConnectViewRenderer } from './device-connect-view-renderer';
