+++ src/electron/views/device-connect-view/device-connect-view.scss
