+++ src/electron/views/device-connect-view/deviceConnectView.html
@@ -6,14 +6,14 @@ Licensed under the MIT License.
 <html dir="ltr" lang="en">
     <head>
         <meta charset="utf-8" />
-        <link rel="stylesheet" type="text/css" href="../../common/styles/fabric.min.css" />
+        <link rel="stylesheet" type="text/css" href="../../../common/styles/fabric.min.css" />
         <link rel="stylesheet" type="text/css" href="./styles/default/device-connect-view.css" />
-        <link rel="stylesheet" type="text/css" href="../../electronBundle/deviceConnectView.css" />
+        <link rel="stylesheet" type="text/css" href="../../../electronBundle/deviceConnectView.css" />
         <title>Accessibility Insights for Mobile</title>
     </head>
 
     <body class="ms-fabric is-focusVisible">
         <div id="device-connect-view-container"></div>
-        <script src="../../bundle/deviceConnectView.bundle.js"></script>
+        <script src="../../../bundle/deviceConnectView.bundle.js"></script>
     </body>
 </html>
