+++ src/tests/unit/tests/electron/device-connect-view/components/device-connect-body.test.tsx
@@ -6,8 +6,8 @@ import {
     DeviceConnectBody,
     DeviceConnectBodyDeps,
     DeviceConnectBodyProps,
-} from 'electron/device-connect-view/components/device-connect-body';
-import { DeviceConnectState } from 'electron/device-connect-view/components/device-connect-state';
+} from 'electron/views/device-connect-view/components/device-connect-body';
+import { DeviceConnectState } from 'electron/views/device-connect-view/components/device-connect-state';
 import { shallow } from 'enzyme';
 import * as React from 'react';
