+++ src/tests/unit/tests/electron/device-connect-view/components/device-connect-header.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DeviceConnectHeader } from 'electron/device-connect-view/components/device-connect-header';
+import { DeviceConnectHeader } from 'electron/views/device-connect-view/components/device-connect-header';
 import { shallow } from 'enzyme';
 import * as React from 'react';
