+++ src/tests/unit/tests/electron/device-connect-view/components/device-connect-port-entry.test.tsx
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-import { DeviceConnectPortEntry, DeviceConnectPortEntryProps } from 'electron/device-connect-view/components/device-connect-port-entry';
+import { DeviceConnectPortEntry, DeviceConnectPortEntryProps } from 'electron/views/device-connect-view/components/device-connect-port-entry';
 import { DeviceConnectActionCreator } from 'electron/flux/action-creator/device-connect-action-creator';
 import { shallow } from 'enzyme';
 import { Button } from 'office-ui-fabric-react/lib/Button';
@@ -8,9 +8,9 @@ import * as React from 'react';
 import { EventStubFactory } from 'tests/unit/common/event-stub-factory';
 import { IMock, Mock, MockBehavior, Times } from 'typemoq';
 
-import { portNumberField } from 'electron/device-connect-view/components/device-connect-port-entry.scss';
+import { portNumberField } from 'electron/views/device-connect-view/components/device-connect-port-entry.scss';
 import { EnumHelper } from '../../../../../../common/enum-helper';
-import { DeviceConnectState } from '../../../../../../electron/device-connect-view/components/device-connect-state';
+import { DeviceConnectState } from '../../../../../../electron/views/device-connect-view/components/device-connect-state';
 
 describe('DeviceConnectPortEntryTest', () => {
     const testPortNumber = 111;
