+++ src/tests/unit/tests/electron/device-connect-view/device-connect-view-renderer.test.tsx
@@ -1,14 +1,14 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import { BrowserWindow } from 'electron';
-import { DeviceConnectViewRenderer } from 'electron/device-connect-view/device-connect-view-renderer';
+import { DeviceConnectViewRenderer } from 'electron/views/device-connect-view/device-connect-view-renderer';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import { IMock, It, Mock } from 'typemoq';
 import {
     DeviceConnectViewContainer,
     DeviceConnectViewContainerProps,
-} from '../../../../../electron/device-connect-view/components/device-connect-view-container';
+} from '../../../../../electron/views/device-connect-view/components/device-connect-view-container';
 
 describe('DeviceConnectViewRendererTest', () => {
     test('render', () => {
